requirements.txt
Django>=4.2.10
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=23.5
django-storages>=1.14.2
djoser>=2.2.2
Pillow>=10.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=22.0.0
trino>=0.327.0
pandas>=2.1.3
numpy>=1.26.1 